Don't kill child processes on restart
This change adds the following to the [Service] section of the
spec:

KillMode=process
SendSIGKILL=no

This will prevent os-refresh-config from being killed when a run
results in os-collect-config upgrading itself.

The original os-refresh-config will continue to conclusion, and
a concurrent os-refresh-config run is prevented with its existing
lockfile protection.

Change-Id: Ib6be3aabd6e7ebfd4fe3e9b78355bfeb0bd7a014
Closes-Bug: #1603144