Update patch set 1
Patch Set 1:

> need to bump the release to get new build in CBS

How much do we care about this? The bug was reported in the TripleO context, which would use the dlrn build.

I would just bump the release, but I need to backport this through several branches and it would a bit more of a hassle, so if I can avoid that it would be nice.

Patch-set: 1
Reviewer: Gerrit User 335 <335@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0